+++ app/build.gradle.kts
@@ -68,6 +68,7 @@ dependencies {
     implementation(project(":bottomnavigation"))
     implementation(project(":core:common"))
     implementation(project(":core:designsystem"))
+    implementation(project(":domain:auth"))
     implementation(libs.constraintlayout)
     implementation(libs.coil.core)
     implementation(libs.flexbox)
