+++ app/src/main/java/com/teampophory/pophory/data/network/model/auth/AuthResponse.kt
@@ -1,7 +1,7 @@
 package com.teampophory.pophory.data.network.model.auth
 
-import com.teampophory.pophory.data.model.auth.Token
-import com.teampophory.pophory.data.model.auth.UserAuthentication
+import com.teampophory.pophory.auth.entity.Token
+import com.teampophory.pophory.auth.entity.UserAuthentication
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
