+++ app/src/main/java/com/teampophory/pophory/feature/auth/model/OAuthToken.kt
@@ -1,15 +1,8 @@
 package com.teampophory.pophory.feature.auth.model
 
+import com.teampophory.pophory.auth.entity.KakaoOAuthToken
+
 typealias KakaoToken = com.kakao.sdk.auth.model.OAuthToken
-
-sealed interface OAuthToken {
-    val accessToken: String
-}
-
-data class KakaoOAuthToken(
-    override val accessToken: String,
-    val refreshToken: String,
-) : OAuthToken
 
 fun KakaoToken.toOAuthToken() = KakaoOAuthToken(
     accessToken = accessToken,
