+++ app/src/main/java/com/teampophory/pophory/feature/onboarding/OnBoardingActivity.kt
@@ -12,10 +12,9 @@ import com.teampophory.pophory.common.context.snackBar
 import com.teampophory.pophory.common.view.viewBinding
 import com.teampophory.pophory.config.di.qualifier.Kakao
 import com.teampophory.pophory.data.local.PophoryDataStore
-import com.teampophory.pophory.data.model.auth.UserAccountState
 import com.teampophory.pophory.databinding.ActivityOnBoardingBinding
-import com.teampophory.pophory.domain.AuthUseCase
-import com.teampophory.pophory.domain.AutoLoginConfigureUseCase
+import com.teampophory.pophory.auth.usecase.AuthUseCase
+import com.teampophory.pophory.auth.usecase.AutoLoginConfigureUseCase
 import com.teampophory.pophory.feature.auth.social.OAuthService
 import com.teampophory.pophory.feature.home.HomeActivity
 import com.teampophory.pophory.feature.onboarding.adapter.OnBoardingViewPagerAdapter
@@ -76,12 +75,12 @@ class OnBoardingActivity : AppCompatActivity() {
                     authUseCase(token.accessToken)
                         .onSuccess { state ->
                             when (state) {
-                                UserAccountState.REGISTERED -> {
+                                com.teampophory.pophory.auth.entity.UserAccountState.REGISTERED -> {
                                     autoLoginConfigureUseCase(true)
                                     startActivity(HomeActivity.getIntent(this@OnBoardingActivity))
                                 }
 
-                                UserAccountState.UNREGISTERED -> {
+                                com.teampophory.pophory.auth.entity.UserAccountState.UNREGISTERED -> {
                                     val intent =
                                         Intent(this@OnBoardingActivity, SignUpActivity::class.java)
                                     startActivity(intent)
