+++ app/src/main/java/com/teampophory/pophory/feature/setting/SettingViewModel.kt
@@ -2,7 +2,7 @@ package com.teampophory.pophory.feature.setting
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.teampophory.pophory.data.repository.auth.AuthRepository
+import com.teampophory.pophory.auth.repository.AuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
