+++ domain/auth/src/main/java/com/teampophory/pophory/auth/entity/OAuthToken.kt
@@ -1,0 +1,10 @@
+package com.teampophory.pophory.auth.entity
+
+sealed interface OAuthToken {
+    val accessToken: String
+}
+
+data class KakaoOAuthToken(
+    override val accessToken: String,
+    val refreshToken: String,
+) : OAuthToken
