+++ domain/auth/src/main/java/com/teampophory/pophory/auth/entity/Token.kt
@@ -1,4 +1,4 @@
-package com.teampophory.pophory.data.model.auth
+package com.teampophory.pophory.auth.entity
 
 data class Token(
     val accessToken: String,
