+++ domain/auth/src/main/java/com/teampophory/pophory/auth/entity/UserAccountState.kt
@@ -1,4 +1,4 @@
-package com.teampophory.pophory.data.model.auth
+package com.teampophory.pophory.auth.entity
 
 enum class UserAccountState(val value: Boolean) {
     REGISTERED(true),
